Add generate.sh script for generating proto descriptors for config.
Add the generate script plumbing for generating proto descriptors for
use with the developing project boxster device configuration
approach. Note that the descriptor file is not checked in as this is the
desired approach to avoid merge conflicts on this frequently changing
binary file.

TEST=./generate.sh to generate a descriptor file
BUG=chromium:1043997

Change-Id: I6ed08bff4889b577f9b199fa9120f17d227c5be3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2013783
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>