Move the starlark/* files to util/, point at local protos.
Move the current starlark/* to the location of the rest of the starlark
helper code in util/ and point proto usage at the locally path mounted
protos.

We never loved the starlark/ name but didn't come up with a better
alternative. Charles pushed back on not using a more descriptive name so
going with util/ for now.

BUG=chromium:1041872
TEST=generating from src/project/volteer/volteer

Change-Id: Ifad6acd5eddad4c49116a14fdc024a08b39e1a86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2056133
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>