Start a README.md limited to instructions for partner checkout for now.
Start a README.md file in this project. For now it includes instructions
on how a partner working on a specific project would get started and
uses the already present local_manifest.xml approach. Not sure if this
is the ideal location for such docs but at least having this laid out
will get us started with our own folks vetting the approach.

BUG=None
TEST=None

Change-Id: I4b191f331fa43e329a5ff34b22c74dd56fd2ace2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2074435
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>