Add a bin/ directory with adding to path config generation approach.
As suggested in https://crrev.com/c/2084931, as opposed to that approach
use an approach that adds a bin/ directory the user will put on the path
and they then invoke configuration from this utility on the path.

I know the name stinks, please advise.

BUG=chromium:1041872
TEST=have tested on a project repo

Change-Id: I46550fb5c3eb9c3f878b6568b78becfab7bf9b64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2085855
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>