config: Add build-target config
Not totally sure where to put this for now.  The entire build system
expects this data from cros_config and merging starlark with some hybrid
yaml or something in the build system doesn't feel that great.

So starting with a bit of build-target metadata for now that will be
published by the respective programs and we'll see how it evolves.

BUG=None
TEST=None

Change-Id: Ic9db697fb245d281bd06208693b643cf4f580d6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2098129
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>