Copy cros_config_proto_converter to config repo, use in gen_config.
Copy the cros_config_proto_converter into the config repo and then use
it within gen_config when we are looking at a project (not a program).

This is part of the migration to checking platform json into each
project repo instead of the current approach where the json is generated
on the fly by the ebuild. The first steps will get this working here,
have projects check in platform json. The last step will be to switch
the ebuild to aggregating the platform json rather than generating it
itself.

BUG=None
TEST=gen_config config.star in src/project/volteer/halvor (reg. project)
TEST=gen_config config.star in src/program/zork (program)
TEST=gen_config config.star in src/program/zork/zork (nested project)

Change-Id: I4e7cf9705f9a0ee89b9c8ecc9bffe5ef6c433a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2123047
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Tested-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>