Make setup_project.sh more resilient to being ran a second time.
Make setup_project.sh more resilient to being ran a second time by
checking for the presence of the directory of the project and if it
already there it deletes it and continues the setup from the start. This
will likely be a common case where the setup fails because the user
fails to have permissions on some of the repos from the local
manifest. In this case they will be able to repair and get it to work
after the permssions are right without having to repair manually or do
the repo init from scratch.

BUG=None
TEST=running again with the project already setup

Change-Id: I0b6b7238e9c1cfd8941a6690ffc0b5e4637140f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2130690
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>