Adjust advised command to regenerate test json expectations.
The old value worked for the script in its old location. This approach
cds to the directory and runs from there. It could also do:

PYTHON_PATH=payload_utils python3 ....

and would not require changing to the directory. Either way it needs to
be able to see the config symlink in payloa_utils.

BUG=None
TEST=ran the command with changes in the json

Change-Id: I66ef0bcda3be8b130dd0a070b2d652aa9ffdb3b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2135159
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>