Split audio codec into headphone and speaker
Boards typically have separate codecs for speakers and headphones,
and while in the past these were usually a fixed pairing with the
volteer project we have more combinations possible.

- Add MAX98373, MAX98357, and MAX98360 speaker amp codecs
- Split 'audio_codec' into 'headphone_codec' and 'speaker_amp' and
update the create_audio() function.  All of these are now optional
parameters in order to aid the transition.
- Update the fake board to use the new prototypes and use ALC5682I
instead of RT5682 so the latter can be removed in a subsequent commit.

BUG=b:147462631
TEST=generate and run tests

Change-Id: If81301206f94e734a7abd210a88a2383d866ae46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2154244
Commit-Queue: Jett Rink <jettrink@chromium.org>
Tested-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>