Remove project path from bluetooth config
This forces a convention with portage that cros_workon must map by
project name, which is unnecessary.

Instead, portage is going to process files on a per project basis, so
the convention is no longer necessary and the path can just be relative
to the project dir in the first place.

BUG=None
TEST=cq

Change-Id: Ifeaadff59683499cfb27a7d4f0c79881144bb8d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2161329
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>