test: Drop support for heterogeneous DUT conditions in metadata
Use of multiple DUTs for sharding test execution is a test plan concept.
The test metadata now simply specifies the conditions that need to be
mett for each DUT that is used for test execution.

Since this API is in alpha stability level, the deprecated field will be
deleted once clients migrate.

BUG=chromium:1069902
TEST=./generate.sh

Change-Id: I60e3e0cb0bc16bf333988b8f433a3d35b3d98a9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2161525
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>