touch test files for many vendors
Currently touch firmware is all based on convention around fw version
naming, so we're going to keep that (at least for now), but remove the
overhead around generating the config.

Instead, once a touch file is added, the config will be automatically
generated based on a shared starlark lib that does the per vendor
specific convention as detected.

This starts that process by committing a collection of test cases for
touch vendors that will be used to develop/test the shared library for
automatically generating the config.

BUG=None
TEST=None

Change-Id: I01d39de77eaad6d029d8c580f9c545c95a962da3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2174886
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Commit-Queue: C Shapiro <shapiroc@chromium.org>