Centralize scripts to "bootstrap" a boxster project in sbin/ directory.
As noted in:

https://docs.google.com/document/d/1_BWvr79Ai044zdAoMBcCOdIiw8vRXBdVZwoA0XbUAU4/edit#

centralizing the scripts that are used to bring up a new boxster
project, only used by googlers, in an sbin/ directory. This separates
them from the bin/ scripts used by partners and the sbin name is similar
to sbin/ in unix where these are admin only scripts.

BUG=None
TEST=None

Change-Id: I3a51cbfa95f38c79f57edff5062247893b274113
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2187918
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>