Have create_project_buckets not fail on non-existent google group.
It appears that sometimes some of the expected google groups that are
applied as objectViewers do not exist. For example the oem and soc
groups appear to occasionally not be created. In such cases allow the
script to continue but output a message that it was unable to add that
group.

BUG=chromium:1081757
TEST=ran with dedede wheelie showing lack of oem and soc groups

Change-Id: Ia7bebc836d621cc2ce565ae174d65a937a3f10b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2197181
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>