cros_config_proto_converter: Add camera_config.json handling.
Add camera_config.json handling to the cros_config_proto_converter.py
script. It looks for files names
sw_build_config/.../camera/camera_config_$DESIGN.json and if found it
wires it up as the config-file field of the camera property.

Note the change from the _set method to _upsert. Since the output under
the camera key is now produced in two different places it requires an
update rather than a set, as with a set the last would clobber the
earlier.

BUG=chromium:1095104
TEST=gen_config ./test/program/fake/config.star && gen_config ./test/project/fake/fake/config.star

Change-Id: I33c3439a379aa1e207340121e66ff5105954837b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2249693
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>