Switch formatting / linting starlark to lucicfg.
Switch the formatting and linting of starlark from using buildifier to
using `lucicfg fmt` and `lucicfg lint`. This also removes the need for
buildifier to be fetched / placed on the path.

Note this is a second attempt with this change - now it appears the
chroot does have a version of lucicfg with the fmt option.

BUG=chromium:1102409
TEST=CQ

Change-Id: I9f6b2bc1123f76ea744e041fb30d5d077c8abe66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2306940
Reviewed-by: Sean McAllister <smcallis@google.com>
Tested-by: David Burger <dburger@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>