gen_config: Trap exit and echo if successful or not.
In the gen_config script trap the exit and echo out whether successful
or not. On more than one occasion a user has put forth CLs that appear
to have gotten past `lucicfg generate` but are failing before
`convert_proto_enums.py`. It seems that there is some failure path that
is not making it obvious that it failed. This will make it obvious.

BUG=None
TEST=locally breaking the script and observing output

Change-Id: Ib03d80815c104afe2b2839433a85095726b90b0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2313657
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>