test: relax fw_config topology tests
We were asserting too strict of a requirement: that the fw_config value
could not be shared by two different hardware topology objects if their
ID's were different. We only need to ensure that their types are the
same, not that that their (type, id) is the same. It is completely valid
to have two different topology objects of the same type with the same
FW_CONFIG value.

BUG=b:163791037
TEST=unit tests

Change-Id: I426fab9adc70ee5fd5651aa4b7537ac63aacbbb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2352424
Tested-by: Jett Rink <jettrink@chromium.org>
Auto-Submit: Jett Rink <jettrink@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>