Finish move to platform config repo
CLs don't depend on each other, but this is the successor to:
crrev.com/c/2359013

dut was already moved to src/config and plan needs to be a platform
versioned construct, which infra depends on, but is not defined or
managed by infra.

This is a step in moving to test planning management that is milestone
versioned under src/config-internal.

Change-Id: Iffef19d7ee2dcd92d06cdf45db4a50b642b3aefa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2358757
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@google.com>
Commit-Queue: Prathmesh Prabhu <pprabhu@google.com>
Auto-Submit: C Shapiro <shapiroc@chromium.org>