Modify build_config script to generate program config too.
Since we depend on it and we're moving to not generating
config.jsonproto as part of the check-in, we need to avoid
race conditions between the program repo and the project
repo being updated.  So just always generate the config for
the program repo on the fly.

BUG=chromium:1108383
TEST=gen_config

Change-Id: I545ec9bc3650185cf46c8c81d29a6ce012b9df6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2378199
Tested-by: Sean McAllister <smcallis@google.com>
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>