Add API field to store SSFC definitions
For a give program, there are two different fields that dictate how
firmware should behave. The FW_CONFIG is constrained by the
DesignConfigID, and it encode the expected functionality of the device
-- whether to use sensors or not.

The other firmware field is the Second Source Factory Cache, and it is
not constrained by the DesignConfigId. The definition of this field must
be set a program level, but it is not set through the configuration
data. The SSFC field is set through factory-time probing.

More information at go/second-source-factory-cache

BUG=b:167938257
TEST=builds

Change-Id: Ide6b5556a3c372f5361f612464af8f9b2122cd91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2406126
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Auto-Submit: Jett Rink <jettrink@chromium.org>