Reviving test_plan util module
This is based on the initial work by jkusuma@ in creating a prototype
impl of the test plan (e.g. chrome-internal:3205695)

Moving most of the util over since it's supporting protos are in this repo
and also removed wifi specific prototyping bits out of the helper, which
will be built out in the prototype impl that uses this helper.

BUG=None
TEST=Tested with internal prototyping config to come

Cq-Depend: chrome-internal:3271755
Change-Id: I2fcc2e1a5ae8288386917d7077e7ab6d9c933d82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2412244
Reviewed-by: Jason Kusuma <jkusuma@google.com>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>