Create design entries for wildcard SKUs as well.
If a device doesn't have a sku-id configured in model.yaml, then
it becomes a 'wildcard' sku that can match any sku.  Currently we
ignore these, but especially in the case of pre-unibuild migrated
boards, there's no actual SKU value available.

To fix this, when we encounter a wildcard SKU, go ahead and create
a design for it, but set the DesignConfigId to {project}:* and
set the firmware_sku value to -1.

BUG=None
TEST=manual

Change-Id: If2846153ee440a3e0d7642de1ea8d809fb267695
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2533444
Reviewed-by: Sean McAllister <smcallis@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>