Allow overriding of platform-name field in platform json.
Right now this value is always set to the program name, but we
need to be able to override it for backwards compatibility with
mosys.  Allow setting it in the program proto and use it when
creating the json rather than the program name.

BUG=None
TEST=manual

Change-Id: I827a80002ac55586d0357fb7d2e4f8c5b7d493d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2537295
Auto-Submit: Sean McAllister <smcallis@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>