tls: FakeOmaha resource and its operations
This CL is a part of the TLS API proposal for fake Omaha service related
operations. Please refer [1] for the proposal content.

In this CL, we define the input/output of TLS APIs to start/stop a fake
Omaha service which run with the TLS.

We also provide examples to show how to use the APIs.

[1] https://docs.google.com/document/d/1xArPqCdvV9pANQ5uwCIhWHOdxi2p1e7uL8qdkEMQkQc/edit#

BUG=None
TEST=Ran `./run_go_unittests.sh`

Change-Id: I88b11ede4385d60e69c211df1b66d1d6194e936a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2500777
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Congbin Guo <guocb@chromium.org>