Connect to AVL database and merge in project form factor.
This is needed to replace the deviceconfig information on
whether a device runs on A/C power or A/C + battery power.
It's really an OR operation on the three form factors that
are desktop units.

Diff on imported data for all current projects:
  https://paste.googleplex.com/6273390169030656?raw

BUG=chromium:1162277
TEST=manual

Change-Id: Ie0d5c7af47ebd2d056724010584b163e06382442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2643982
Commit-Queue: Sean McAllister <smcallis@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>