Connect to AVL database and merge in project form factor.
This is needed to replace the deviceconfig information on
whether a device runs on A/C power or A/C + battery power.
It's really an OR operation on the three form factors that
are desktop units.

Diff on imported data for all current projects:
  https://paste.googleplex.com/6273390169030656?raw

BUG=chromium:1162277
TEST=manual

Change-Id: Ic834bd967f15eb5f2fef8908dd296db46fac024d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2691306
Commit-Queue: Sean McAllister <smcallis@google.com>
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Auto-Submit: Sean McAllister <smcallis@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>