Modify join_config_payloads to populate designs properly.
We can really have design information come from three sources:
  * config.jsonproto payload (specified in starlark)
  * model.yaml file
  * specified as part of the script (--project-name)

To get consistent results, we need to reconcile these three
sources and makes sure we end up with a design created for
all of them.  This is done by aggregating/deduplicating the
names and then forcing the designs to exist (even if they're empty).

The one sharp edge here is that when specifing a --config-bundle
we now need to specify whether we're only using it as a reference
for the models (--import-only).

uber diffs:
  imported: https://paste.googleplex.com/4774280723169280?raw
  joined: https://paste.googleplex.com/5628054857580544?raw

BUG=chromium:1183413
TEST=uber diff

Change-Id: Ie853ea3ed0d6486fab48608b2c947cce369ec337
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2779794
Commit-Queue: Sean McAllister <smcallis@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>