config: Add support for custom web apps dir
We need to be able to customize the set of default installed web apps in
Chrome per device model. In model.yaml files, it is already possible to
specify a "extra web apps dir", which gets transformed into an ash flag
at build time (see crrev.com/c/2750846).

This CL adds the same mechanism to the new style of Starlark
configuration. Individual configs can specify a directory name, which
will be converted into an ash flag by cros_config_proto_converter.

This new configuration is added to a new 'ui_config' proto, which is
intended to roughly match the structure of the fields in model.yaml.

BUG=chromium:1182481
TEST=./run_py_unittests.sh

Change-Id: I8a8c0df81b9c1f64e849bfb793b8351c95873d23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2810579
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>