boxster: modify gen_config to take --no-proto option
It's good to regenerate the protos usually, but if you want to run
a bunch of gen_config in parallel, then the step of removing the protos
and then regenerating causes them to step on each other, so let's
just add a workaround to disable the regeneration step when we know
that we don't need it.

Default behavior is unchanged.

BUG=None
TEST=manual

Change-Id: I964ea52479e5207cdc79862f82e148ec33dc696a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2831841
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>