Initial DutService definition
This is an excise/clone of the ExecDutCommand and FetchCrashes api calls
from commontls.proto.

This is going to be a stand alone service since it has general
utility/use outside of the scope of test execution.

E.g. lab maint and provisioning ops depend only on this service, but not
on test execution (with which is much heavier weight).

Also, this will support device state capture, which can be run in or out
of scope of a test execution.  E.g. lab ops will can capture device
state after a reboot that crashes, etc... (outside the scope of
testing).

BUG=None
TEST=./generate

Change-Id: Id8e09d0fbfec8153c21574cbb0c2386bdfa86b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/2920275
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>