Initial inventory service for getting dut topology
Starting the breakup of tls/tlw into smaller, decoupled, and more easily
implemted services for the local-tle, el.all tles.

This brings in first-class support for multi-dut topologies also as the
defacto unit that will be scheduled and provided all the way down to the
test harnesses.

BUG=b:188712103
TEST=./generate.sh

Change-Id: Iee02c603bf3b405842f43d071ccd744f0b8dad6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3001082
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>