Making identity scan config a hw fn
Making this a fn of cros_config was a problem in that it's not a
property of the software (since this config is set during manufacturing
and is immutable for the life of the device).

Having it as software config made it unaccessible to dut management
services that need to probe for identity.

This also tries to make it more obvious/intuitive with naming and
putting the scan config with the ids they map back to.

This is a non-breaking proto change ... it's a pure message rename with
all contents/types remaining completing compatible.

BUG=None
TEST=generate & cq

Change-Id: I8fbcef3e90a6b8f1b1a62b2f66a59b2d2e4100f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3001096
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>