Decoupling designconfig from mfgconfig
Besides making this ambigous/confusing, this is problematic because the
data comes from different places.

This keeps is separated, so clients (test-runner, ...) can choose if
they need to ask for this data or not from the respective tle.

For local, we may opt to not support mfg-config initially also.

BUG=None
TEST=generate

Change-Id: I7206461d6c0fa6367db312d9d76b331f4d09ca65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3001098
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>