Lab-env only must return device config ids
All of this config will be available as a build artifact for CI builds
and locally in the src tree for local test execution, so it will be a
shared service that actually retrieves the config data.

The lab will only be responsible for returning the ids for the lookup.
In the local lab case, it will scan for the ids via dut_service, but in
the managed lab, they will be stored/provided to the drone tle.

BUG=b:188712103
TEST=generate

Change-Id: If9f5b8efa343af4fe029026cb552692d96f7fc20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3035389
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>