Break out provisioning namespace
The OWNERS enforcement here is off, so beginning the process of fixing
it by breaking out the appropriate directories (on top of the cl that
fixes all of the gen'd dirs so OWNERS is auto-included there).

This allows AU team to own the service def going forward after the
migration is completed.

BUG=None
TEST=cq

Change-Id: I43cb390dd45381e31aae31b9cc3ffecb276e20fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3093287
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>