release_target: nest variants and add criticality
We should keep variants as full fledged nested versions of the release
proto. This keeps a hierarchy but allows them to have their own uniq
configuration w.r.t. signing, build_target, and criticality. It's
expected we won't go more than one variant deep but we might design to
allow it.

Criticality is added with a bit more granularity as an enum so that we
can more finely control the builder's impact on LKGM, dashboarding
visibility and the overall return status of releases.

BUG=b:199540054
TEST=regenerated

Change-Id: Ie11a02483891578c117257dd825151002a69a6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3161316
Reviewed-by: Sean McAllister <smcallis@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>