dut-attributes: Add aliases and remove software labels
Add aliases for board and model to help with confusing nomenclature,
and remove the software attributes.  sw-build-target in particular is
a build property so it doesn't make sense to target it for scheduling.

The latest qualified firmware version makes a little more sense, but
I don't know that we schedule on it now so let's remove it until we're
sure people need it (before we're committed to keeping it).

BUG=none
TEST=manual

Skip-Dut-Attrs-Check: OK to remove, not live yet
Change-Id: Ieaf3e9fcbac72ed51e05a7cc309f2f423666479e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3170741
Auto-Submit: Sean McAllister <smcallis@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Justin Suen <justinsuen@google.com>
Commit-Queue: Sean McAllister <smcallis@google.com>