test-api: Add ProvisionConfig and wrap into new HWTestPlan message.
We need to specify some provisioning config for testing, notably
the firmware versions to flash and any DLC content to install on the
device.

It doesn't make sense to expose the full ProvisionState to the end user
since there's settings in there that they shouldn't/can't modify.

BUG=b:196070598
TEST=./generate.sh

Change-Id: Ic8e95e9426381574d5d742b197c224dfc41ca8ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3176391
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>