test-api: Move test plan to versioned folder.
This is so we can be explicit if we ever have to make a breaking
change.  The CTP2 API will take a oneof{} of the test plan versions
and handle the conversion to a ToT format.

BUG=b:202415435
TEST=./generate.sh

Change-Id: Iaa6db84a78af2513367d9359ccdc49fba473a1ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3212667
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>
Commit-Queue: Sean McAllister <smcallis@google.com>
Commit-Queue: Tim Bain <tbain@google.com>