dut-attributes: add TleSource attributes
Add lab-defined attributes along with their aliases. The labels are
defined in go/f20-dutattribute-map. The values of these attributes are
to be looked up by the TLE caller. Each caller is responsible for
providing the implementation to map each attribute ID to its
corresponding value in the TLE.

BUG=b/204117838
TEST=./generate; presubmit

Change-Id: I293fd33ee051996d1b8f579c4013f0d673aad46b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3277305
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>