Add a brand config field for gaming devices, generating a ui flag.
Features to be enabled only on gaming devices will be gated by a
corresponding feature flag. These may be only for a subset of design
SKUs, but not necessarily be tied to any particular hardware properties.
As such, configure this in brand config, allowing a particular design to
always opt-in, by enabling gaming_device for their BrandConfig, or for
a selection of devices, controllable via additional BrandConfigs using
custom label or rebranding functionality.

BUG=b:231387993
TEST=./run_py_unittests.sh

Change-Id: Idff51756c82a8f602c5398f01b64ee111b9e7101
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/3627420
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Michael Martis <martis@chromium.org>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>