Rename NOT_PRESENT to UNKNOWN for Cellular type
Currently in Boxster organization, there is a present flag in the Cellular component. However, in the Cellular type we also have an enum value for NOT_PRESENT. This is duplicated information and can easily be set for contradictory situations, for example, the present flag is true but the type says NOT_PRESENT.

Therefore to change NOT_PRESENT to CELLULAR_UNKNOWN to indicate an unknown type, which is better in turn of usage.

BUG=b:270564386
TEST=./generate.sh and gen_config with fake project

Change-Id: I8bf88abd8637dc3a15a85ca0c5b2142747d55d24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4285508
Commit-Queue: YH Lin <yueherngl@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>