Stop checking that fake project is regenerated in PRESUBMIT.
- Since proto converter is in platform2, changes to proto
converter can cause diffs in the fake project and break these
presubmits. Stop checking this in the presubmit, since it is not
critical that the fake project is up to date. Don't completely
remove the fake project since some of the unit tests in this
repo use it as test data.

BUG=b:270394023
TEST=CQ

Change-Id: I5f794adbe6be37b45bfea4a218b1c7290b55492a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4288570
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>