vmlab: update CTRv2 cacheserver template to support vm test use case
VM test in Test Runner v2 utilize the cache server container to mimic the cache server in hardware test. To authenticate the cache server, the template asks for a service account keyfile to be mounted and exported as an env var. The auth field is a oneof to support future use case of running locally where a config folder should be mounted instead.

BUG=b:272318078
TEST=./generate.sh

Change-Id: I1da8726ae28d0d7d3e8b56740fe6830478118b3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4337987
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Ming Kong <mingkong@google.com>