Map recovery input to keyboard for detachables
The recovery_input field is used to confirm physical presence when
enabling developer mode in firmware recovery screen. The value
POWER_BUTTON was designed for Wilco devices (sarien, drallion), where
their EC is not trusted. For detachables, although the Power button is
used in practice to confirm physical presence, the field should be set
to KEYBOARD (we should probably rename it to DEFAULT to avoid
confusion). Otherwise the instructions in the firmware screen would be
incorrect.

BUG=b:275033700
TEST=gen_config

Change-Id: I6997b55bed188ade1d8b26d329f89880a2d9f69f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4369384
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yidi Lin <yidilin@chromium.org>
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>