Updating AndroidOsImage proto with os_version field.
This CL introduces os_version field in the AndroidOsImage proto. It can be used instead of gs_path. The OS version will be mapped to the Android image location by Android provision service as it knows DUT model and board.

This is a non-breaking and fully backward compatible change.

Change-Id: Ib484566d2eeeba5a2c7048075480566eb4c17967
Bug=b:275112828
TEST=./generate.sh
TEST=bin/gen_config test/project/fake/fake/config.star
TEST=./run_py_unittests.sh
TEST=./run_go_unittests.sh
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4382790
Commit-Queue: Sergey Fetisov <sfetisov@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>