config: add spi_flash_transform field to project config
The factory configuration needs to have a string to string mapping that
lets them tranform an output flash chip name from `futility flash`
command into the input flash name for the `ap_wpsr` tool. Most of the
time this mapping will be empty, but there are cases where we will need
to go from a general "flash name" to a more specific flash name within
the context of a project. See go/config-updates-for-ap-ro-verification
for more background.

BUG=b:233971354
TEST=generate.sh passes
TEST=added new field to fake and see correct generated output

Change-Id: I29979464c4b41a2ce13b22dfe554b7853bb224fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4520734
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>