proto: Add bluetooth_peer proto
Adds a new bluetooth_peer.proto for btpeer related messages. This
includes a new BluetoothPeerChameleondConfig message which is an
adaptation of the current yaml config managed manually. This will
be used to read and modify the GCS config file that manages the
expected Chameleond versions btpeers should have based on chameleond
commit and dut hostname and version. See the message and field
comments for more details.

BUG=b:282836502
TEST=generate.sh ran successfully

Change-Id: I44b1932f337fcfe7428586501077c90db9c5e526
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4700389
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Jared Bennett <jaredbennett@google.com>