vm-leaser: implement SubnetModeNetworkEnabled for GCP configs
Some GCP networks have subnet mode enabled and require subnets to be
passed when creating GCE VM instances. This field lets VM Leaser check
if that mode is enabled so that it can get the appropriate zone subnet
value and use it in the lease request.

BUG=b/283819096
TEST=./generate.sh

Change-Id: I06fff1a445a3a008d04bc1844d8f50df851d6820
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4639918
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Deniz Kara <denizkara@google.com>