vm-leaser: add ListLeases RPC to the VM Leaser API
This RPC follows the guidelines of AIP-132 (https://google.aip.dev/132)
for List requests and AIP-160 (https://google.aip.dev/160) for
filtering. The first iteration of the RPC will only implement filtering
for metadata and tags, allowing users to traverse those fields for
values to match with.

The protos are not annotated as suggested by AIP because
google.api.*.proto are not supported in this repo currently.

BUG=b/279053406
TEST=./generate.sh

Change-Id: I20e8c98d3f1a8171b35c90908cdcccd34bd645a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4744194
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Deniz Kara <denizkara@google.com>