Add ResourceConfig as a platform and SoC topology config.
ResourceConfig is often set once per SoC platform so in most cases a
single good default can be set at the program level. In cases where it
needs to diverge from the platform-level defaults, allow resource config
to be provided in the SoC topology, overlaying the config onto the
platform-level config. Move the resource config proto out of software so
it can be reused at the program/platform level and in topologies.

BUG=b:301198820
TEST=./run_py_unittests.sh

Cq-Depend: chromium:4886430
Change-Id: Ib8d3fe9262c708068917c109789f97249b80b5ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4886070
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>